first push.
this is the base game with all the bugs and problems. 
most of the keybord code is made by deepseek r1.
im to stupped to understand usb controlers and input devices.

i hope you find the progect intresting.

build command:
gcc -O2 -lm -o out rock.c &&sudo ./out

i am using nixos so that is the reason for the -lm to include the math librery. if you use windows or other distro you coud remove it. (i guess)

the rock.h is the space ship assets. have it in the same dir when conpiling.

run the program ass sudo so it can access the keybord controler.

;3